requirements.txt
langchain
langgraph
openai
faiss-cpu
sentence-transformers
python-dotenv
pydantic